pipeline/pyproject.toml
[project]
name = "tacit-knowledge-pipeline"
version = "0.1.0"
description = "Tacit knowledge mining pipeline: ingest sources → extract norms → detect contradictions → generate questions"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.40",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
docling = ["docling>=2.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
